vpn_detector/requirements.txt
pandas
numpy
pyarrow
scikit-learn>=1.3
xgboost
matplotlib
seaborn
pyyaml
joblib
pyshark
tqdm
